requirements.txt
colorama==0.4.4
fastapi==0.63.0
loguru==0.5.3
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.2
pydantic==1.8.1
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
starlette==0.13.6
torch==1.8.0
torchvision==0.9.0
typing-extensions==3.7.4.3
win32-setctime==1.0.3
pytorch-lightning==1.4.5
matplotlib
seaborn
sklearn
jupyter
tqdm
opencv-python
albumentations